refactor(hush-cli): extract trust from pkg_cli
Move the registry trust-verification cluster (RegistryAttestation,
RegistryProof, AttestationVerification, checkpoint_signature_message,
verify_attestation_against_hash, verify_checkpoint_signature,
verify_transparency_proof, InstalledIdentity,
required_registry_public_key_for_trust, verify_install_trust) into
pkg_cli/trust.rs. Items consumed by the install + verify handlers
widened fn/struct -> pub(super); internal-only helpers stay private.
Struct fields constructed by tests/install made pub(super).

mod.rs imports for LeafData and checkpoint_signature_message are now
test-only (their non-test consumer moved with the cluster), so they
carry #[cfg(test)]; merkle/crypto imports unused after the move were
trimmed.
